Verification-pass fixes: no-extras collection, TCP-reset retry, test teeth
Fixes surfaced by an adversarial review of the previous commit:

- conftest: spec.py is back on the no-extras ignore list — it imports lazily,
  but its doctests CONSTRUCT http/physionet sources, which trips the extras
  ImportError at example-execution time and would have redded the
  "core (no extras)" CI matrix. In exchange, the fsspec-only download tests
  (CLI end-to-end + failure exits + key validation + collision rejection +
  memory:// round-trip) move to a new httpx-free tests/test_download_fsspec.py,
  so the no-extras job now exercises the base-install download path end-to-end
  instead of skipping the download surface entirely. Verified against a real
  `uv sync --locked --group dev` environment (93 passed).
- http: a mid-body TCP reset surfaces as httpx.ReadError, which was in neither
  retry set — one RST killed the file despite the new Range-resume retry.
  ReadError/WriteError join _RETRY_EXC, and the resume regression test is
  parametrized over both mid-body failure shapes.
- The gzip-server regression test was vacuous (transparent decoding also
  produces correct bytes on a fresh download); rewritten as a resume scenario
  against an RFC-compliant gzip-encoding server, which fails without the
  Accept-Encoding: identity fix.
- CLI/README key-guard docs now state the no-sources-block carve-out (warn +
  exit 0) instead of overclaiming "always an error".
- CLI failure tests pin their exit to the intended error message; the
  memory:// test uses a unique root and cleans up the process-global
  filesystem; the SIGINT docstrings no longer claim daemon-thread teardown.
