+++ conftest.py
@@ -9,10 +9,13 @@ The ``collect_ignore_glob`` block skips the HTTP-backed download modules (+ thei
 doctests) when the ``download`` extra (``httpx``, ``tenacity``) isn't installed. Those
 two modules raise ``ImportError`` at import time when the extras are missing, which is
 the right library behavior but trips up pytest's ``--doctest-modules`` collector that
-imports every module unconditionally. The rest of the download layer (``source.py``,
-``spec.py``, ``cli.py``, the fsspec backend) imports lazily and stays collected. The
-sibling ``run_tests_download`` CI job installs the extra and runs only the download
-tests; see ``.github/workflows/tests.yaml``.
+imports every module unconditionally; ``spec.py`` imports fine but its doctests
+construct http/physionet sources, which trips the same ImportError at example time.
+The rest of the download layer (``source.py``, ``cli.py``, the fsspec backend, and
+``tests/test_download_fsspec.py``) stays collected, so the no-extras job exercises
+the fsspec-only download path end-to-end. The sibling ``run_tests_download`` CI job
+installs the extra and runs the full download surface; see
+``.github/workflows/tests.yaml``.
 """
 
 from __future__ import annotations
@@ -50,6 +53,9 @@ except ImportError:
         [
             "src/MEDS_extract/download/backends/http.py",
             "src/MEDS_extract/download/backends/physionet.py",
+            # Imports lazily, but its doctests CONSTRUCT http/physionet sources,
+            # which triggers the extras ImportError at example-execution time.
+            "src/MEDS_extract/download/spec.py",
             # Imports httpx at module top for its MockTransport-based tests.
             "tests/test_download.py",
         ]
